Fix "TypeError: write() argument must be str, not bytes" on Python 3
This used to happen when writing to stdout. On Python 3 stdout is an
instance of TextIOWrapper which is a text-based interface. It wraps a
bytes-based writer though which is accessible through the buffer
attribute, so let's use that to either write to it directly or to get an
encoding writer.

Fixes https://github.com/vilcans/screenplain/issues/53.